requirements/base.txt
aiohttp==3.9.5
aiosignal==1.3.1
alembic==1.13.1
async-timeout==4.0.3
asyncpg==0.29.0
attrs==23.2.0
certifi==2024.2.2
charset-normalizer==3.3.2
exceptiongroup==1.2.1
frozenlist==1.4.1
gino==1.0.1
idna==3.7
iniconfig==2.0.0
Mako==1.3.5
MarkupSafe==2.1.5
multidict==6.0.5
packaging==24.0
pluggy==1.5.0
psycopg2==2.9.9
pytest==8.2.1
python-dotenv==1.0.1
requests==2.32.3
SQLAlchemy==1.3.24
tomli==2.0.1
typing_extensions==4.12.0
urllib3==2.2.1
uvloop==0.19.0
yarl==1.9.4
